backend/requirements.txt
bcrypt==3.2.0
fastapi==0.63.0
google==3.0.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
psycopg2==2.8.6
SQLAlchemy==1.4.2
uvicorn==0.13.4
python-telegram-bot==13.5
pyjwt==2.1.0
pydantic~=1.8.2
starlette~=0.13.6